170516 메세지팝업4개 구현 / PopupActivity 인스턴스변수추가 / msg Time(long타입변환)
* 메세지팝업 4개 구현
친구요청메세지 팝업 구현
음성메세지 팝업 구현
기분메세지 팝업 구현
진동메세지 팝업 구현
->테스트는 제대로 못해봤다

* popupAct.java (String ID, long Time 추가)

* MsgDAO, MsgDTO : Time타입 int에서 long으로 변환

+++ HeartBeat_App/HeartBeat/app/src/main/java/comjungwon7769heartbeat/github/heartbeat/MsgDAO.java
@@ -59,7 +59,7 @@ public class MsgDAO extends SQLiteOpenHelper {
 	}
 
 	//DELETE Message
-	public boolean deleteMsg(String sender, int time) {
+	public boolean deleteMsg(String sender, long time) {
 		SQLiteDatabase db = this.getWritableDatabase();
 
 		String sql = "DELETE FROM " + Table_name + " WHERE " + Sender + " = '" + sender + "' AND " + Time + "=" + time;
@@ -111,7 +111,7 @@ public class MsgDAO extends SQLiteOpenHelper {
 	}
 
 	//Msg 반환 Method
-	public MsgDTO getMsg(String sender, int time) {
+	public MsgDTO getMsg(String sender, long time) {
 		SQLiteDatabase db = this.getReadableDatabase();
 		Cursor cursor = db.rawQuery("SELECT * FROM " + Table_name + " WHERE " + Sender + "='" + sender + " AND " + Time + " = " + time, null);
 
@@ -141,7 +141,7 @@ public class MsgDAO extends SQLiteOpenHelper {
 	}
 
 	//진동 카운트 없데이트 Method
-	public void updateMsg(String sender, int flag, int count, int time) {
+	public void updateMsg(String sender, int flag, int count, long time) {
 		SQLiteDatabase db = this.getWritableDatabase();
 
 		if(flag != Constants.msgFlag_Bzz) {
